Linux/gcc: fix -Wsign-compare across the tree and enable the flag
Resolve all 69 signed/unsigned comparison sites surfaced by -Wsign-compare and
stop suppressing it in CMakeLists (kept on as a ratchet; tree is clean under it).

All fixes are behaviour-preserving for the actual runtime value ranges:
- sentinel checks `unsigned == -1` made explicit as `== T(-1)` / UNDEFINED_FILE_SIZE
  (bit-identical: the int -1 already converted to the all-bits-set unsigned value).
- non-negative loop indices retyped to the matching unsigned type, or cast up at
  the comparison; non-negative pointer/iterator differences cast up to SizeT.
- GridCoord backward scan: made the already-implicit signed->unsigned conversion
  explicit (an off-grid negative maps to a large value that differs from the
  in-range currGridPos, so the loop still breaks).

Touched: quotes, FileMapHandle, FileView.h, RingIterator.h, SharedStr, FixedAlloc,
TriggerOperator (rtc); gdal_vect/gdal_grid/gdal_base, BmpImp, ShpImp, TifImp(.h),
GridStorageManager.h, XdbImp, Odbc{Imp,StorageManager} (stg); Modus, CalcClassBreaks,
OperAttrBin, GetStatistics (clc); SelCaret, FeatureLayer, GridCoord, ItemSchemaView
(shv); OperPolygon, BoostPolygon (geo); SessionData (tic).

+++ rtc/dll/src/act/TriggerOperator.cpp
@@ -175,7 +175,7 @@ void DecRemainingTargetCount()
 
 	void LeaveNotificationBlock()
 	{
-		dms_assert(g_NrTargetsAtEnter != -1);
+		dms_assert(g_NrTargetsAtEnter != UInt32(-1));
 		if (g_NrTargetsAtEnter != g_RemainingTargetCount)
 		{
 			leveled_critical_section::scoped_lock notifyLock(sc_NotifyTargetCount);
